refactor(repository): fixes imports
